контроллер для идентификации юзера

+++ src/main/java/com/amr/project/dao/PersonalDataRepository.java
@@ -3,5 +3,5 @@ package com.amr.project.dao;
 import com.amr.project.model.entity.PersonalData;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PersonalDataRepository extends JpaRepository<PersonalData, Integer> {
+public interface PersonalDataRepository extends JpaRepository<PersonalData, Long> {
 }
